Move context logic to it's file and abstract the provider to make consumer code cleaner.
